Get rid of clever setting of domain member
It matched IP addresses, which was the exact opposite of what it was supposed to do.

In order for verification to proceed properly, we'll just set domain to whatever host is set to -- even if it's an IP address.

If it proceeds to verification without being overridden by use of the `-domain` flag, the TLS lib is smart enough to check for IP SANs.